requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
tiktoken==0.5.1
openai>=1.50.0
python-dotenv==1.0.0
pandas==2.1.0
streamlit==1.28.0
plotly==5.17.0
numpy==1.24.3
httpx>=0.24.0
